// src/graphql/orders/mutations/__generated__/orderPrint.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: OrderPrint
// ====================================================

export interface OrderPrint {
  orderPrint: string;
}

export interface OrderPrintVariables {
  storeId: string;
  orderId: string;
}